requirements.txt
streamlit
pandas
numpy
sqlalchemy
pymysql
python-dotenv
scikit-learn
scipy
cryptography
